DataTypeChangeFromIntToLong

+++ UVP.ExternalIntegration.Business/Interfaces/IIntegrationOrchestrationService.cs
@@ -2,7 +2,7 @@
 {
     public interface IIntegrationOrchestrationService
     {
-        Task<bool> ExecuteFullClearanceCycleAsync(int doaCandidateId, int candidateId, string integrationType);
-        Task<bool> CheckAndProgressClearanceAsync(int doaCandidateId, int candidateId, string integrationType);
+        Task<bool> ExecuteFullClearanceCycleAsync(long doaCandidateId, long candidateId, string integrationType);
+        Task<bool> CheckAndProgressClearanceAsync(long doaCandidateId, long candidateId, string integrationType);
     }
 }
