+++ UVP.ExternalIntegration.Business/Interfaces/IInvocationManagerService.cs
@@ -4,7 +4,7 @@ namespace UVP.ExternalIntegration.Business.Interfaces
 {
     public interface IInvocationManagerService
     {
-        Task<long> CreateInvocationAsync(int doaCandidateId, int candidateId, string integrationType);
+        Task<long> CreateInvocationAsync(long doaCandidateId, long candidateId, string integrationType);
         Task<long> CreateInvocationAsync(IntegrationRequestDto request);
         Task<bool> ProcessPendingInvocationsAsync();
         Task<bool> ProcessRetryableInvocationsAsync();
