+++ UVP.ExternalIntegration.Business/ResultMapper/DTOs/ResultMappingContext.cs
@@ -2,8 +2,8 @@
 {
     public class ResultMappingContext
     {
-        public int DoaCandidateId { get; set; }
-        public int CandidateId { get; set; }
+        public long DoaCandidateId { get; set; }
+        public long CandidateId { get; set; }
         public string IntegrationType { get; set; } = string.Empty;
         public string Operation { get; set; } = string.Empty;
         public string Response { get; set; } = string.Empty;
