+++ UVP.ExternalIntegration.Domain/DTOs/IntegrationRequestDto.cs
@@ -2,8 +2,8 @@
 {
     public class IntegrationRequestDto
     {
-        public int DoaCandidateId { get; set; }
-        public int CandidateId { get; set; }
+        public long DoaCandidateId { get; set; }
+        public long CandidateId { get; set; }
         public string IntegrationType { get; set; } = string.Empty;
         public string? IntegrationOperation { get; set; }
     }
