+++ UVP.ExternalIntegration.Domain/Entities/DoaCandidate.cs
@@ -2,9 +2,9 @@
 {
     public class DoaCandidate
     {
-        public int Id { get; set; }
-        public int DoaId { get; set; }
-        public int CandidateId { get; set; }
+        public long Id { get; set; }
+        public long DoaId { get; set; }
+        public long CandidateId { get; set; }
         public string? Department { get; set; }
         public string? RequestorName { get; set; }
         public string? RequestorEmail { get; set; }
